+++ admin/mail/database.php
@@ -1,5 +1,7 @@
 <?php
 
+if ( ! isset($CFG) ) exit;
+
 $DATABASE_UNINSTALL = array(
 "drop table if exists {$CFG->dbprefix}mail_send",
 "drop table if exists {$CFG->dbprefix}mail_bulk");
